- Used Numeric data type everywhere in db layer - Used BigDecimal everywhere in app layer - Manual string to big decimal calculation from the CSV - Used only two crypto for testing BTC and ETH

+++ backend/src/main/java/com/prem/ta/dtos/MarketDataDTO.java
@@ -38,9 +38,9 @@ public record MarketDataDTO(
         BigDecimal volumeProfileVAL,
 
         @JsonProperty("bvs")
-        Double buyerVolumeShare,
+        BigDecimal buyerVolumeShare,
 
         @JsonProperty("bcs")
-        Double buyerCapitalShare
+        BigDecimal buyerCapitalShare
 ) {
 }
